==> 1_MergeDataset.ipynb <==
import os, warnings, random
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import csv

from google.colab import drive
drive.mount('/content/drive',force_remount=True)

data_folder = '/content/drive/MyDrive/master-iot-data/dataset'
output_file = 'merged_data.csv'
attack_types = {
    'EVSE-A-charging-Aggressive-scan.csv': 'Aggressive Scan',
    'EVSE-A-charging-icmp-flood.csv': 'ICMP Flood',
    'EVSE-A-charging-icmp-fragmentation.csv': 'ICMP Fragmentation',
    'EVSE-A-charging-os-fingerprinting.csv': 'OS Fingerprinting',
    'EVSE-A-charging-portscan.csv': 'Port Scan',
    'EVSE-A-charging-push-ack-flood.csv': 'Push ACK Flood',
    'EVSE-A-charging-service-detection.csv': 'Service Detection',
    'EVSE-A-charging-slowLoris-scan.csv': 'SlowLoris Scan',
    'EVSE-A-charging-syn-flood.csv': 'SYN Flood',
    'EVSE-A-charging-syn-stealth.csv': 'SYN Stealth',
    'EVSE-A-charging-synonymous-ip.csv': 'Synonymous IP',
    'EVSE-A-charging-tcp-flood.csv': 'TCP Flood',
    'EVSE-A-charging-udp-flood.csv': 'UDP Flood',
    'EVSE-A-charging-vulnerability-scan.csv': 'Vulnerability Scan',
    'EVSE-A-idle-aggressive-scan.csv': 'Aggressive Scan',
    'EVSE-A-idle-benign.csv': 'Benign',
    'EVSE-A-idle-icmp-fragmentation.csv': 'ICMP Fragmentation',
    'EVSE-A-idle-os-fingerprinting.csv': 'OS Fingerprinting',
    'EVSE-A-idle-portscan.csv': 'Port Scan',
    'EVSE-A-idle-service-detection.csv': 'Service Detection',
    'EVSE-A-idle-slowloris-scan.csv': 'SlowLoris Scan',
    'EVSE-A-idle-syn-flood.csv': 'SYN Flood',
    'EVSE-A-idle-syn-stealth-scan.csv': 'SYN Stealth',
    'EVSE-A-idle-synonymous-ip.csv': 'Synonymous IP',
    'EVSE-A-idle-tcp-flood.csv': 'TCP Flood',
    'EVSE-A-idle-udp-flood.csv': 'UDP Flood',
    'EVSE-A-idle-vulnerability-scan.csv': 'Vulnerability Scan'
}

# Tüm dosyaları birleştirme işlemi
merged_data = pd.DataFrame()
for file_name, attack_type in attack_types.items():
    file_path = os.path.join(data_folder, file_name)
    data = pd.read_csv(file_path)
    data['Attack Type'] = attack_type
    merged_data = pd.concat([merged_data, data], ignore_index=True)

# Çıktıyı kaydetme
output_file = os.path.join(data_folder, 'merged_data.csv')
merged_data.to_csv(output_file, index=False)
print("Veriler birleştirildi ve {} dosyasına kaydedildi.".format(output_file))

dataset =pd.read_csv('/content/drive/MyDrive/master-iot-data/dataset/merged_data.csv')
data = dataset.copy()
data
