==> 3_Preprocessing.ipynb <==
import os, warnings, random
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import csv

from google.colab import drive
drive.mount('/content/drive',force_remount=True)

dataset =pd.read_csv('/content/drive/MyDrive/master-iot-data/dataset/merged_data.csv')
data = dataset.copy()
data

print("Gözlem sayısı : " ,len(data))
print("Değişken sayısı : ", len(data.columns))

data['Attack Type'].value_counts()

NaN ve Inf Değerleri Kontrol Etme


#NaN değere sahip sutünları kaldırma
data.drop(["requested_server_name","client_fingerprint","server_fingerprint","user_agent","content_type"],axis=1,inplace=True)


non_integer_columns = data.select_dtypes(exclude=['int']).columns
print("Non-integer columns:", non_integer_columns)

data['bidirectional_mean_ps'] = data['bidirectional_mean_ps'].round().astype(int)
data['bidirectional_stddev_ps'] = data['bidirectional_stddev_ps'].round().astype(int)
data['src2dst_mean_ps'] = data['src2dst_mean_ps'].round().astype(int)
data['src2dst_stddev_ps'] = data['src2dst_stddev_ps'].round().astype(int)
data['dst2src_mean_ps'] = data['dst2src_mean_ps'].round().astype(int)
data['dst2src_stddev_ps'] = data['dst2src_stddev_ps'].round().astype(int)
data['bidirectional_mean_piat_ms'] = data['bidirectional_mean_piat_ms'].round().astype(int)
data['bidirectional_stddev_piat_ms'] = data['bidirectional_stddev_piat_ms'].round().astype(int)
data['src2dst_mean_piat_ms'] = data['src2dst_mean_piat_ms'].round().astype(int)
data['src2dst_stddev_piat_ms'] = data['src2dst_stddev_piat_ms'].round().astype(int)
data['dst2src_mean_piat_ms'] = data['dst2src_mean_piat_ms'].round().astype(int)
data['dst2src_stddev_piat_ms'] = data['dst2src_stddev_piat_ms'].round().astype(int)

data['Attack Type'].value_counts()

attack_types = data['Attack Type'].unique()
print(attack_types)

group = { 'Benign': 0,
          'SYN Flood': 1,
          'Synonymous IP': 2,
          'TCP Flood': 3,
          'Push ACK Flood': 4,
          'UDP Flood': 5,
          'SYN Stealth': 6,
          'Port Scan': 7,
          'Vulnerability Scan': 8,
          'Aggressive Scan': 9,
          'OS Fingerprinting': 10,
          'Service Detection': 11,
          'SlowLoris Scan': 12,
          'ICMP Fragmentation': 13,
          'ICMP Flood': 14}



data['Attack_Num'] = dataset['Attack Type'].map(lambda x: group[x])
data['Attack_Num'].value_counts()

# Attack_Num sütununda değeri 13 veya 14 olan satırları seçin
rows_to_drop = data[(data['Attack_Num'] == 13) | (data['Attack_Num'] == 14)   ].index

# Seçilen satırları veri çerçevesinden silin
data.drop(rows_to_drop, inplace=True)

# Silme işleminden sonra veri çerçevesini yeniden indeksleyin
data.reset_index(drop=True, inplace=True)


data['Attack_Num'].value_counts()

#Label_Num = Label , Cat_Num = Cat , Sub_Cat_Num = Sub_Cat
data.drop(["Attack Type"],axis=1,inplace=True)

columns_to_drop = ['id','expiration_id', 'src_port' , 'dst_port','src_ip', 'dst_ip', 'src_mac', 'dst_mac', 'src_oui', 'dst_oui', 'application_category_name' , 'application_name']
data.drop(columns=columns_to_drop, inplace=True)

non_integer_columns = data.select_dtypes(exclude=['int']).columns
print("Non-integer columns:", non_integer_columns)

data.to_csv("/content/drive/MyDrive/master-iot-data/cleaned_merged_data.csv", index = False)

from sklearn.preprocessing import MinMaxScaler

# Min-Max normalizasyonu için bir ölçekleyici oluşturun
scaler = MinMaxScaler()

# 'atack_num' sütununu dışındaki veri sütunlarını ölçeklendirin
scaled_data = data.copy()  # Veriyi korumak için kopyalayın
columns_to_scale = scaled_data.columns.difference(['Attack_Num'])  # 'atack_num' sütunu dışındaki sütunları seçin
scaled_data[columns_to_scale] = scaler.fit_transform(scaled_data[columns_to_scale])  # Seçilen sütunları ölçeklendirin

# Normalleştirilmiş veriyi görüntüleyin
print(scaled_data.head())
data = scaled_data.copy()


data.to_csv("/content/drive/MyDrive/master-iot-data/cleaned_merged_data_MinMax.csv", index = False)
