==> 4_Feature_Selection.ipynb <==
import os, warnings, random
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import csv

from google.colab import drive
drive.mount('/content/drive',force_remount=True)

# dataset =pd.read_csv('/content/drive/MyDrive/master-iot-data/cleaned_merged_data.csv.csv')
dataset =pd.read_csv('/content/drive/MyDrive/master-iot-data/cleaned_merged_data_MinMax.csv')
data = dataset.copy()
data

threshold = 0.01  # Eşik değeri belirleyin

# Varyansı hesaplayın
variances = data.var()

# Eşik değerden düşük olan sütunları seçin
selected_features = variances[variances > threshold].index

# Seçilen özellikleri görüntüleyin
print("Selected Features:")
print(selected_features)

# Seçilen özellikleri içeren yeni bir veri kümesi oluşturun
selected_data = data[selected_features]


data = selected_data.copy()

# Tüm özelliklerin adlarını alın
all_features = data.columns

# Seçilen özellikleri içeren listeden tüm özelliklerin adlarını çıkarın
not_selected_features = set(all_features) - set(selected_features)

# Seçilmeyen özellikleri görüntüleyin
print("Not Selected Features:")
print(not_selected_features)


data.to_csv("/content/drive/MyDrive/master-iot-data/feature_selection_data.csv", index = False)
