==> 5_SMOTE.ipynb <==
import os, warnings, random
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import csv

from google.colab import drive
drive.mount('/content/drive',force_remount=True)

# dataset =pd.read_csv('/content/drive/MyDrive/master-iot-data/cleaned_merged_data.csv.csv')
#dataset =pd.read_csv('/content/drive/MyDrive/master-iot-data/cleaned_merged_data_MinMax.csv')
dataset =pd.read_csv('/content/drive/MyDrive/master-iot-data/feature_selection_data.csv')
data = dataset.copy()
data

from imblearn.over_sampling import SMOTE


# Bağımsız değişkenler ve hedef değişkeni ayırın
X = data.drop(columns=['Attack_Num'])  # Bağımsız değişkenler
y = data['Attack_Num']  # Hedef değişken

# SMOTE örnekleme uygulayın
smote = SMOTE(k_neighbors=5, sampling_strategy='auto')

X_resampled, y_resampled = smote.fit_resample(X, y)

# Yeni örnekleme ile güncellenmiş veri setini oluşturun
resampled_data = pd.concat([pd.DataFrame(X_resampled, columns=X.columns), pd.DataFrame(y_resampled, columns=['Attack_Num'])], axis=1)

# Sonuçları göster
print(resampled_data.head())

data = resampled_data.copy()


data.to_csv("/content/drive/MyDrive/master-iot-data/SMOTE_data.csv", index = False)
